backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.0.3

# QR Code generation
qrcode[pil]==7.4.2
pillow==10.1.0

# HTTP client for external APIs
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Redis for caching (commented out for now)
# redis==5.0.1
# aioredis==2.0.1